Create SettingViewModelHelper - Move the creation and containment of all Setting VM classes to this new class. - Remove isVisible from SettingBase and children. - Make the formerly invisible ViewModel properties into properties that use UserSettings properties directly. - The SettingViewModels underlying collection was never actually observed so converted into a List.
