+++ TechtonicaModLoader/MVVM/ViewModels/Settings/SettingBase.cs
@@ -1,9 +1,4 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace TechtonicaModLoader.MVVM.ViewModels.Settings
 {
@@ -14,15 +9,18 @@ namespace TechtonicaModLoader.MVVM.ViewModels.Settings
         public string Name { get; }
         public string Description { get; }
         public string Category { get; }
-        public bool IsVisible { get; }
 
         // Constructors
 
-        public SettingBase(string name, string description, string category, bool isVisible) {
+        public SettingBase(string name, string description, string category) {
             Name = name;
             Description = description;
             Category = category;
-            IsVisible = isVisible;
+
+            if (!Description.EndsWith('.')) {
+                Description += '.';
+            }
+
         }
     }
 }
